backend/requirements.txt
fastapi
uvicorn
opencv-python-headless
pillow
imagehash
sqlalchemy
python-multipart
numpy
scikit-image

